feat: extract cmp
